removed dependency/import of delayed.js in blocks
