+++ scripts/delayed.js
@@ -1,42 +1,11 @@
 // eslint-disable-next-line import/no-cycle
-import { sampleRUM, getMetadata, fetchPlaceholders } from './lib-franklin.js';
+import { sampleRUM, getMetadata } from './lib-franklin.js';
 
 // Core Web Vitals RUM collection
 sampleRUM('cwv');
 
 // add more delayed functionality here
 
-// eslint-disable-next-line import/prefer-default-export
-export async function isMobile() {
-  const mql = window.matchMedia('(max-width: 600px)');
-
-  return mql.matches;
-}
-
-export function formPictureTag(pictureClass, mobileImgUrl, desktopImgUrl) {
-  const picture = document.createElement('picture');
-  picture.className = pictureClass;
-
-  const sourceDesktop = document.createElement('source');
-  sourceDesktop.media = '(min-width: 768px)';
-  sourceDesktop.srcset = desktopImgUrl;
-  sourceDesktop.type = 'image/webp';
-  picture.appendChild(sourceDesktop);
-
-  const sourceMobile = document.createElement('source');
-  sourceMobile.media = '(max-width: 767px)';
-  sourceDesktop.type = 'image/webp';
-  sourceMobile.srcset = mobileImgUrl;
-  picture.appendChild(sourceMobile);
-
-  const img = document.createElement('img');
-  img.src = desktopImgUrl;
-  img.alt = '';
-  img.width = '1024';
-  img.height = '750';
-  picture.appendChild(img);
-  return picture;
-}
 // add serves and duration to recipe details page content
 function AddServesAndDuration() {
   this.content = (serves, duration) => {
@@ -138,37 +107,6 @@ function UpdateExternalLinks() {
   };
 }
 new UpdateExternalLinks().init();
-
-export function ProcessStoriesBgImage() {
-  this.updateStoriesBgImage = async (params) => {
-    const placeholder = await fetchPlaceholders();
-    const storyContainer = document.querySelector('main');
-
-    const createAndAppendPicture = (bgClass, bgConstant) => {
-      if (document.querySelector(`.${bgClass}`)) document.querySelector(`.${bgClass}`).remove();
-      return formPictureTag(bgClass, placeholder[`${bgConstant}Mobile`], placeholder[`${bgConstant}Desktop`]);
-    };
-
-    const topPicture = createAndAppendPicture(params.BG_TOP_CLASS, params.BG_TOP);
-    storyContainer.insertBefore(topPicture, storyContainer.firstChild);
-
-    const bottomPicture = createAndAppendPicture(params.BG_BOTTOM_CLASS, params.BG_BOTTOM);
-    storyContainer.appendChild(bottomPicture);
-  };
-  this.init = () => {
-    if (document.querySelector('body.story-tips')) {
-      const bgConfigParams = {
-        BG_TOP: 'storyTipsBgTop',
-        BG_BOTTOM: 'storyTipsBgBottom',
-        BG_TOP_CLASS: 'story-page-bg-top',
-        BG_BOTTOM_CLASS: 'story-page-bg-bottom',
-      };
-      const boundAction = this.updateStoriesBgImage.bind(this, bgConfigParams);
-      boundAction();
-    }
-  };
-}
-new ProcessStoriesBgImage().init();
 
 /**
  * Google Tag Manager implementation
